feat: implement TT tables for move ordering

+++ src/movefinding/board.h
@@ -62,8 +62,8 @@ typedef struct Position_t
     bool white_to_move;
     uint16_t num_children;
     uint16_t half_move_count;
-    int64_t piece_value_diff;
-    int64_t evaluation;
+    int32_t piece_value_diff;
+    int32_t evaluation;
     ULL all_pieces;
     ULL en_passant_bitboard;
     ULL zobrist_key;
